// src/redux/company/clients/constants.ts
export enum ClientTypes {
    API_RESPONSE_SUCCESS = '@@company/clients/API_RESPONSE_SUCCESS',
    API_RESPONSE_ERROR = '@@company/clients/API_RESPONSE_ERROR',

    GET_CLIENTS = '@@company/clients/GET_CLIENTS',
    GET_CLIENT = '@@company/clients/GET_CLIENT',
    ARCHIVE_CLIENT = '@@company/clients/ARCHIVE_CLIENT',
    RESET = '@@company/clients/RESET',
}